+++ src/main/java/com/eodegano/cocobackend/repository/MstSigunguRepository.java
@@ -1,0 +1,7 @@
+package com.eodegano.cocobackend.repository;
+
+import com.eodegano.cocobackend.domain.MstSigungu;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MstSigunguRepository extends JpaRepository<MstSigungu, String> {
+}